requirements.txt
numpy>=1.19.5
h5py>=2.10.0
Pillow>=8.0.0